@@ -11,15 +11,17 @@
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "directories": {},
+  "type": "module",
+  "sideEffects": false,
   "engines": {
-    "node": ">=14"
+    "node": ">=14.13.1 || >=16.0.0"
   },
   "files": [
     "dist"
   ],
   "scripts": {
     "build": "rimraf dist && tsc -p tsconfig.prod.json",
-    "test": "jest",
+    "test": "NODE_OPTIONS=--experimental-vm-modules jest",
     "cz": "git-cz"
   },
   "repository": {
@@ -42,11 +44,13 @@
     "openapi-json-schema": "^1.1.3"
   },
   "devDependencies": {
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-typescript": "^7.18.6",
     "@types/jest": "^29.4.0",
     "cz-conventional-changelog": "^3.3.0",
     "jest": "^29.4.1",
     "rimraf": "^4.1.2",
-    "ts-jest": "^29.0.5",
+    "ts-jest-resolver": "^2.0.0",
     "typescript": "^4.9.5"
   },
   "config": {
