@@ -1,8 +1,9 @@
-module.exports = {
-	preset: 'ts-jest',
+export default {
+	resolver: 'ts-jest-resolver',
 	testEnvironment: 'node',
 	testMatch: [ '<rootDir>/lib/**/*.spec.ts' ],
 	collectCoverage: true,
 	collectCoverageFrom: [ '<rootDir>/lib/**' ],
 	coverageReporters: [ 'lcov', 'text', 'html' ],
+	extensionsToTreatAsEsm: ['.ts'],
 };
